refactor/#11: Move MemberPrincipalDetail
