+++ src/main/java/com/picpic/server/room/service/CreateRoomService.java
@@ -1,29 +1,31 @@
 package com.picpic.server.room.service;
 
-import com.picpic.server.common.security.MemberPrincipalDetail;
+import org.springframework.stereotype.Service;
+
+import com.picpic.server.common.auth.MemberPrincipalDetail;
 import com.picpic.server.common.util.IdUtils;
 import com.picpic.server.room.service.usecase.CreateRoomUseCase;
 import com.picpic.server.room.service.usecase.RedisRoomCommandUseCase;
 import com.picpic.server.room.service.usecase.RoomHistoryCommandUseCase;
+
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class CreateRoomService implements CreateRoomUseCase {
 
-    private final RedisRoomCommandUseCase redisRoomCommandUseCase;
-    private final RoomHistoryCommandUseCase roomHistoryCommandUseCase;
+	private final RedisRoomCommandUseCase redisRoomCommandUseCase;
+	private final RoomHistoryCommandUseCase roomHistoryCommandUseCase;
 
-    @Override
-    public String createRoom(MemberPrincipalDetail creatorPrinciple, Integer roomCapacity) {
-        String roomId = IdUtils.generateRoomId();
+	@Override
+	public String createRoom(MemberPrincipalDetail creatorPrinciple, Integer roomCapacity) {
+		String roomId = IdUtils.generateRoomId();
 
-        roomHistoryCommandUseCase.create(IdUtils.generateTsid(), creatorPrinciple.memberId());
-        redisRoomCommandUseCase.create(roomId, creatorPrinciple, roomCapacity);
+		roomHistoryCommandUseCase.create(IdUtils.generateTsid(), creatorPrinciple.memberId());
+		redisRoomCommandUseCase.create(roomId, creatorPrinciple, roomCapacity);
 
-        return roomId;
-    }
+		return roomId;
+	}
 }
