+++ src/main/java/com/picpic/server/room/service/usecase/CreateRoomUseCase.java
@@ -1,7 +1,7 @@
 package com.picpic.server.room.service.usecase;
 
-import com.picpic.server.common.security.MemberPrincipalDetail;
+import com.picpic.server.common.auth.MemberPrincipalDetail;
 
 public interface CreateRoomUseCase {
-    String createRoom(MemberPrincipalDetail creatorPrinciple, Integer roomCapacity);
+	String createRoom(MemberPrincipalDetail creatorPrinciple, Integer roomCapacity);
 }
